- Big change to the array system; - TArray is now the base array type and is pure. The allocator policy functionality is now implemented with virtual functions (Mostly GetAllocSize, SetAllocSize) so derivatives dictate allocation. The former heap-allocated array is now THeapArray. - The sort policy system is now gone (from the client perspective) in favour of all client-side comparisons for sorting and searching implemented with < operators on the types. The actual searching is now done with "SortIterators" (these are functoids) which implement either a binary chop or a simple array-walk iteration. The correct iterator is instigated by the SortPolicy. There are two kinds of sort policies... sorted and none. The sorted sort policy also instigates an Insert-Sort after an element has been added. (This is optional on the SortPolicy Template, but certainly required in most cases and is significantly faster than instigating a quicksort on a worst-cast-scenario all-sorted array). There are additional sortpolicy implementations for pointers and Ptr's as they require special interpretation because of the language as well as dereferencing null pointers. - All TArray<TYPE>'s are now interchangeable whereas before they had to have matching policies to be of the same type. - BUT, for some reason the code cannot determine that all arrays could be cast down in the TBinary Read/Write overloads so explicit HeapArray and FixedArray overloads have to be implemented... look out for this if other types are Read/Write'd. - The new system means we handle an array search of any type, and the array code will internally use a binary chop when possible. This also allows two arrays of the same type to be sorted by different types. To aid this distinction, all comparisons are now done with a TSorter. There are default global < operators to handle most default cases. The rendertargets stored in the screens are good examples where the one type (TRenderTarget) can be sorted by TRef or by their Z order. Again, all that is required is the storage definition (which dictates which type to search/sort by) and a specific < and == operator in the type. - The special arrays (PtrArray, PointerArray) now have a base array type to derive from as a template parameter. - Implemented = operators and copy constructors for arrays which was causing problems in the old implementation. Fine now! - The graphs no longer store TPtr nodes and instead just c-pointer nodes. This has been done for speed and gives us an incredible speed up. All graph requests and arrays of nodes (children, node index) are now reallocated with memcpys. Again, massive speed ups! - To go with this, all the zone node stuff is done with c-pointers now instead of smart pointers. - In order to get the code to compile with new global operators I have updated GCC to 4.2... which means the SDK is now required to be 3.0 and above. The xcconfig has changed so 3.0 is now the minimum. - Added TLAsset::GetAllAssets to compliment LoadAllAssets which gets us a list of possible assets for a type, though there could be errors because of files which export to multiple types. - Sped up scheme importing by pre-allocating data and traversing the child node data instead of using TBinaryTree::GetChildren which did a large amount of unneccesary allocations in exceptional cases. - Added a print to TPublisher::DoPublishMessage when the number of subscribers has changed. I'm not sure if this is neccessary, I was using it to catch a subscriber being DELETED but still looping through. Left in for now, but I can't imagine we would NEED to remove it. - Sped up the TRef generation a little and removed the RefSort function. - Allowed a trailing f in TString::GetFloat - In IPodOpenglCanvas, I've changed the code which generates a TRef from a pointer to a more unique ref system. At lower frame rates (I presume) I was getting many breaks where ref's were being duplicated when Touches were made. Presumably the old system wasn't finding them correctly?. Not sure if the new system is fully fixed as I'm not sure if the touch is working perfectly on my ipod, but it doesn't Debug_Break() anymore... - re-organised wx code to be easier to build and stuff, but we may be moving over to QT for tools and UI generation anyway. Changed the code so a wx config can be used better without fiddling with the project. - Added initial QT implementations (including the project xcconfigs) - The sprite render node now defaults to the quad asset.

+++ Code/TootleGui/Mac/MacView.mm
@@ -382,8 +382,8 @@ void releaseData(void *info, const void *data, size_t dataSize)
 -(NSImage*)createImageFromView 
 {
 	NSRect rect = [[NSScreen mainScreen] frame];
-	int width = rect.size.width;
-	int height = rect.size.height;
+	int width = (int)rect.size.width;
+	int height = (int)rect.size.height;
 	
 	NSInteger myDataLength = width * height * 4;
 	GLubyte *buffer = (GLubyte *) malloc(myDataLength);
